requirements.txt
pynput
sounddevice
soundfile
openai-whisper
pyperclip
pyautogui
